# test(consensus): Output e2e latency in the consensus performance test