I'm using `replica_http_ingress_watcher_wait_for_certification_duration_seconds` metrics to compute, essentially doing the following query:
```
avg(delta(replica_http_ingress_watcher_wait_for_certification_duration_seconds_sum[15m]) / delta(replica_http_ingress_watcher_wait_for_certification_duration_seconds_count[15m]))
```

```
2024-09-05 15:44:34.488 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG: Task setup               PASSED  in 132.36s   -- Exited with code 0.
2024-09-05 15:44:34.488 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG: Task test_small_messages PASSED  in 103.57s
2024-09-05 15:44:34.488 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Success rate: 100.0%
2024-09-05 15:44:34.488 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Block rate: 0.4 blocks/s
2024-09-05 15:44:34.601 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Throughput: 1.3 MiB/s, 305.1 messages/s
2024-09-05 15:44:34.601 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Avarage E2E ingress message latency: 14.7s
2024-09-05 15:44:34.601 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG: Task test_large_messages PASSED  in 120.63s
2024-09-05 15:44:34.601 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Success rate: 100.0%
2024-09-05 15:44:34.601 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Block rate: 0.5 blocks/s
2024-09-05 15:44:34.601 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Throughput: 1.3 MiB/s, 1.4 messages/s
2024-09-05 15:44:34.601 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG:      Avarage E2E ingress message latency: 15.4s
2024-09-05 15:44:34.601 INFO[uvms_logs_stream:StdOut] [uvm=test-driver] TEST_LOG: ===========================================================================
```