@kpop-dfinity 2024-09-06
> Nice, I didn't know about that metric

Yeah, me neither, I learned it yesterday from the networking team