@tthebst 2024-09-06
We only added this metric in the last release.